Changes to the ansible playbook to work with a virtual environment. Important to note that this assumes the username on the device is "RPi"!!
Previously the playbook attempted to work with the systemwide python installation. However, in newer versions of Debian, and hence also its derivatives like Ubuntu and Raspberry Pi OS, installing into the systemwide python installation is discouraged (as it risks breaking the entire system since these operating systems depend on python for part of their functionality). This can be circumvented by working with a separate dedicated virtual environment in a new folder that handles all installations. Here it is called 'installations' and is located in the root of user's home directory. This approach has been shown to work, but the only downside being that the directory created by the playbook can't be scp'ed into.

This is a problem as pickle requires all files used to instantiate objects should also be present in the place where unpickling happens ("~/installations/.venv/bin/." in this case), and in the same hierarchy. If this is not satisfied, unpickling will be impossible. In practice this means that the "fission" directory should be present as well as any folders that contain implementation specifics about jobs, pipes, potential utils files etc.